% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Guanine and Cytosine Content of a Sequence}
\usage{
gc_content(sequence)
}
\arguments{
\item{sequence}{A string of characters}
}
\value{
A vector containing the percentage of guanine and cytosine in the sequence
}
\description{
Guanine and Cytosine Content of a Sequence
}
\examples{
gc_content("ATCG")
gc_content("ATTC")
gc_content("ATGG")
}
